Phase 10: Comprehensive tests, PHPStan level 8, GitHub Actions CI
- Added 19 new test files (93 new tests, 142 total)
- Unit tests for all visitors, fixers, analyzers, diff, test generator
- Integration tests for analyzer and fixer pipelines
- 3 fixture files for PHP 5.6, 7.4, and function extraction
- PHPStan level 8 clean (39 errors fixed: type annotations, null safety)
- Fixed DynamicPropertyVisitor: AllowDynamicProperties FQ name support
- Fixed DynamicPropertyVisitor: isset() guard for namespacedName
- GitHub Actions CI: PHP 8.1-8.4 matrix + PHPStan job

+++ src/Diff/DiffResult.php
@@ -16,14 +16,15 @@ final readonly class DiffResult
 
     public function getAddedLineCount(): int
     {
-        return preg_match_all('/^\+[^+]/m', $this->unifiedDiff);
+        return (int) preg_match_all('/^\+[^+]/m', $this->unifiedDiff);
     }
 
     public function getRemovedLineCount(): int
     {
-        return preg_match_all('/^-[^-]/m', $this->unifiedDiff);
+        return (int) preg_match_all('/^-[^-]/m', $this->unifiedDiff);
     }
 
+    /** @return array<string, mixed> */
     public function toArray(): array
     {
         return [
